requirements.txt
fastapi==0.91.0
uvicorn==0.21.1
numpy==1.24.2
pandas==1.5.0
keras==2.9.0
tensorflow==2.9.0
opencv-python-headless==4.5.5.64
Pillow==9.2.0
git-lfs==1.6
